Fixed multiple problems with collisions

+++ src/reference/Astar.h
@@ -21,9 +21,9 @@ template <typename Node>
 template <typename Graph, typename Node>
 std::optional<Path<Node>> a_star_search(Graph &graph, Node start, Node goal) {
   Path<Node> came_from(graph.size());
-  std::vector<std::optional<double>> cost_so_far(graph.size(), std::nullopt);
+  std::vector<std::optional<float>> cost_so_far(graph.size(), std::nullopt);
 
-  PriorityQueue<Node, double> frontier;
+  PriorityQueue<Node, float> frontier;
   frontier.put(start, 0);
 
   came_from[start] = start;
@@ -37,11 +37,11 @@ std::optional<Path<Node>> a_star_search(Graph &graph, Node start, Node goal) {
 
     // For each neighbour
     for (Node s : graph.neighbours(current)) {
-      double new_cost = cost_so_far[current].value() + graph.cost(current, s);
+      float new_cost = cost_so_far[current].value() + graph.cost(current, s);
 
       if (!cost_so_far[s] || new_cost < cost_so_far[s].value()) {
         cost_so_far[s] = new_cost;
-        double priority = new_cost + graph.distance(s, goal);
+        float priority = new_cost + graph.distance(s, goal);
         frontier.put(s, priority);
         came_from[s] = current;
       }
